requirements.txt
aiofiles==0.7.0
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.1
colorama==0.4.4
ecdsa==0.17.0
fastapi==0.68.0
greenlet==1.1.1
h11==0.12.0
idna==3.3
model-dotenv==0.0.0.1
parse==1.19.0
psycopg2-binary==2.9.1
pyasn1==0.4.8
pydantic==1.8.2
pyhumps==3.0.2
python-dotenv==0.14.0
python-jose==3.3.0
python-keycloak==0.26.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.23
starlette==0.14.2
typing-extensions==3.10.0.0
urllib3==1.26.7
uvicorn==0.15.0
